--- front/src/components/Buttons/ButtonIconRound.tsx
@@ -1,10 +1,0 @@
-import { HiOutlinePlus } from 'react-icons/hi'
-import { styles } from '@/styles/components/Buttons/ButtonIconRound.styles'
-
-export const ButtonIconRound = () => {
-  return (
-    <button css={styles.button}>
-      <HiOutlinePlus size={24} />
-    </button>
-  )
-}
